feat(nav): split activiteit into 5-tab navbar + explore placeholder
Phase 2 of Feed/Follow feature. Bottom navbar goes from 3 to 5 tabs
(home stays centered for symmetry).

Structure:
- NEW: geschiedenis.tsx (history list, was activiteit Page 0)
- NEW: rekening.tsx (bills + settlements, was activiteit Page 1)
- NEW: explore.tsx (placeholder for upcoming feed UI)
- DEL: activiteit.tsx (segmented pager no longer needed)

Icons: time / wallet / home / chatbubbles / compass
Aurora palettes sourced from theme.auroraPalettes (history/settlement/explore)

CustomNavBar:
- TAB_COUNT 3 -> 5, PILL_W 80 -> 56, SIDE_MARGIN 24 -> 20
- Programmatic pill/icon positions, no hardcoded arrays
- Active-fallback index adjusted (home now at 2)
